agent: eliminate event creation for l1
